Refactor models into separate files for better organization
- Split `CustomUser`, `Client`, and `Package` models into individual module files.
- Update `models/__init__.py` to import and expose models.
