+++ nbxdjango/packagehandling/models/__init__.py
@@ -1,0 +1,3 @@
+from .user import CustomUser
+from .client import Client
+from .package import Package
