+++ nbxdjango/packagehandling/models/client.py
@@ -1,0 +1,30 @@
+from django.db import models
+from django.conf import settings
+
+class Client(models.Model):
+    user = models.OneToOneField(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.SET_NULL,
+        null=True,
+        blank=True
+    )
+    first_name = models.CharField(max_length=255)
+    last_name = models.CharField(max_length=255)
+    email = models.EmailField()
+    identification_number = models.CharField(max_length=255)
+    state = models.CharField(max_length=255)
+    city = models.CharField(max_length=255)
+    main_street = models.CharField(max_length=255)
+    secondary_street = models.CharField(max_length=255)
+    building_number = models.CharField(max_length=255)
+    mobile_phone_number = models.CharField(max_length=255)
+    phone_number = models.CharField(max_length=255)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    @property
+    def full_name(self):
+        return f'{self.first_name} {self.last_name}'
+
+    def __str__(self):
+        return self.full_name
